requirements.txt
appdirs==1.4.3
cachetools==2.0.0
cloudinary==1.7.0
dill==0.2.6
dj-database-url==0.4.2
Django==1.11
django-cloudinary-storage==0.2.1
django-extra-fields==0.9
django-filter==1.0.2
djangorestframework==3.6.2
editdistance==0.3.1
enum34==1.1.6
future==0.16.0
gapic-google-cloud-vision-v1==0.90.3
google-auth==1.0.0
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-vision==0.24.0
google-gax==0.15.8
googleapis-common-protos==1.5.2
grpcio==1.3.0
gunicorn==19.7.1
httplib2==0.10.3
oauth2client==3.0.0
olefile==0.44
packaging==16.8
Pillow==4.1.1
ply==3.8
proto-google-cloud-vision-v1==0.90.3
protobuf==3.2.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.21
